serviceability: exchange.bgp_community u32 -> u16 (#1929)
## Summary of Changes
* In the serviceability smartcontract, changed exchange.bgp_community
from u32 to u16 since this value represents the right half of a 32-bit
community, where the left half represents a 16-bit autonomous system
number
* This change prepares exchange.bgp_community for use by an upcoming
change to intra-exchange routing policy

## Testing Verification
* Tests updated as needed

+++ smartcontract/cli/src/exchange/create.rs
@@ -23,7 +23,7 @@ pub struct CreateExchangeCliCommand {
     pub lng: f64,
     /// Optional BGP community for the exchange
     #[arg(long)]
-    pub bgp_community: Option<u32>,
+    pub bgp_community: Option<u16>,
 }
 
 impl CreateExchangeCliCommand {
